+++ btcd/wire/msggetnondirdata.go
@@ -97,7 +97,7 @@ func (msg *MsgGetNonDirData) Command() string {
 // receiver.  This is part of the Message interface implementation.
 func (msg *MsgGetNonDirData) MaxPayloadLength(pver uint32) uint32 {
 	// Num inventory vectors (varInt) + max allowed inventory vectors.
-	return MaxVarIntPayload + (MaxInvPerMsg * maxInvVectPayload)
+	return uint32(MaxVarIntPayload + (MaxInvPerMsg * maxInvVectPayload))
 }
 
 // NewMsgGetNonDirData returns a new factom get non dir data message that conforms to the
